fix: move sample to stories file
